Import model in the Menu
