+++ src/app/js/admin/Appbar/ModelMenu.js
@@ -1,18 +1,15 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import { Button, Snackbar } from '@material-ui/core';
+import { Button } from '@material-ui/core';
 import { Add as AddNewIcon } from '@material-ui/icons';
 import { makeStyles } from '@material-ui/core/styles';
-import Alert from '@material-ui/lab/Alert';
 import { connect } from 'react-redux';
 import compose from 'recompose/compose';
 import translate from 'redux-polyglot/translate';
 import { withRouter } from 'react-router';
 
 import AddFromDatasetIcon from './AddFromDatasetIcon';
-import ImportFieldsDialog from './ImportFieldsDialog';
 import { polyglot as polyglotPropTypes } from '../../propTypes';
-import { importFieldsClosed as importFieldsClosedAction } from '../import';
 import { addField } from '../../fields';
 import { showAddColumns, hideAddColumns } from '../parsing';
 
@@ -35,31 +32,12 @@ const useStyles = makeStyles({
 });
 
 export const ModelMenuComponent = ({
-    importFieldsClosed,
-    hasPublishedDataset,
     handleAddColumn,
     handleShowExistingColumns,
     handleHideExistingColumns,
     p: polyglot,
 }) => {
     const classes = useStyles();
-    const [open, setOpen] = useState(false);
-    const [
-        showImportFieldsConfirmation,
-        setShowImportFieldsConfirmation,
-    ] = useState(false);
-    const [showSuccessAlert, setShowSuccessAlert] = useState(false);
-
-    const handleImportFieldsClose = sucess => {
-        importFieldsClosed();
-        setShowImportFieldsConfirmation(false);
-        setShowSuccessAlert(sucess === true);
-    };
-
-    const handleImportFields = () => {
-        setOpen(false);
-        setShowImportFieldsConfirmation(true);
-    };
 
     const handleAddNewColumn = () => {
         handleHideExistingColumns();
@@ -68,14 +46,6 @@ export const ModelMenuComponent = ({
 
     return (
         <div className={classes.container}>
-            <Button
-                variant="outlined"
-                className="btn-import-fields"
-                onClick={handleImportFields}
-                color="primary"
-            >
-                {polyglot.t('import_fields')}
-            </Button>
             <Button
                 variant="contained"
                 color="primary"
@@ -94,18 +64,6 @@ export const ModelMenuComponent = ({
                 <AddNewIcon className={classes.icon} />
                 {polyglot.t('new_field')}
             </Button>
-            {!hasPublishedDataset && showImportFieldsConfirmation && (
-                <ImportFieldsDialog onClose={handleImportFieldsClose} />
-            )}
-            <Snackbar
-                open={showSuccessAlert}
-                autoHideDuration={60000}
-                onClose={() => setShowSuccessAlert(false)}
-            >
-                <Alert variant="filled" severity="success">
-                    {polyglot.t('model_imported_with_success')}
-                </Alert>
-            </Snackbar>
         </div>
     );
 };
@@ -114,12 +72,10 @@ ModelMenuComponent.propTypes = {
     handleAddColumn: PropTypes.func.isRequired,
     handleShowExistingColumns: PropTypes.func.isRequired,
     handleHideExistingColumns: PropTypes.func.isRequired,
-    hasPublishedDataset: PropTypes.bool.isRequired,
     p: polyglotPropTypes.isRequired,
 };
 
 const mapDispatchToProps = {
-    importFieldsClosed: importFieldsClosedAction,
     handleAddColumn: name => addField({ name }),
     handleShowExistingColumns: showAddColumns,
     handleHideExistingColumns: hideAddColumns,
